refactor: extract L2 table and mapping types into meta/l2.rs
Move L2Entry, Mapping, MappingSource, L2Table, their trait impls and
the test_l2_table unit test into a new meta/l2.rs submodule. The test
moves with the code because it reads L2Table's private cluster_bits
field. No logic changes.
